Make sure field settings won't get overwritten
The new 2.3 Duplicator directly displays instances that only have a single template. So when XMLImporter only hides fields when switching sections all sections will still be posted and those with a single template always send an empty result. Because each section starts with the index `1`, previous section will be overwritten.

Using `.detach()` makes sure this won't be a problem.